fix alt arrival calculation, 0.5m/s error due to cast error fix #706 regression commit 092a0044f47a76d306cf58b86bd367246a3e2980

+++ Common/Source/Calc/McReady.cpp
@@ -499,7 +499,7 @@ double GlidePolar::MacCreadyAltitude_internal(double emcready,
 
   #endif // BCT_ALT_FIX
 
-  BestSinkRate = SinkRateFast(0,(int)BestSpeed);
+  BestSinkRate = SinkRateFast(0,BestSpeed);
 
   if (TimeToGo) {
     *TimeToGo = TimeToDestTotal;
